fix: address task switcher review feedback
